+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - *Breaking change*: `clientSecret` is now a named argument in `Client` constructor
 - Add optional `forceRefresh` argument to `getTokenResponse`
 - Add `fromJson` and `toJson` to `Credential` 
+- Add `revoke` method to `Credential`
+
 
 ## 0.2.5